refactor(dto): 당첨 통계를 DTO로 분류
당첨 통계를 나타내는 데이터로 쓰이던 Statistics를 도메인이 아닌 DTO로 분류합니다.
